Final Project - Create a Game
"The Gauntlet" is a continuous, top-down, hockey-themed game where the player controls a skater navigating through incoming defenders while collecting pucks and scoring goals.

Defenders, pucks, and nets spawn at the top and move downward towards the player. The player must collect pucks (max 3) and score goals by skating it into the net. 
The player loses a puck if they score or have 3 pucks and try to collect another (2 pucks remain).
The player loses a life when they collide head-on with a defender, or skate into a net without carrying a puck.

+10 points for scoring (also +1 life if 1 life remaining). 
+1 point for body-checking a defender (by hitting them from the side).

The game gets progressively harder as the speed gradually increases. The player has 3 lives; losing all lives ends the game. The final score is based on the total points scored and the time survived.